move examples to tests; remove makefile
