+++ test/examples/erc20/contracts/ERC20.sol
