+++ test/examples/simulate/simulate.test.ts
@@ -1,6 +1,6 @@
 import expect from 'expect';
 import { Contract } from '../../../src';
-import { loadContract } from '../../utils';
+import { loadContract } from '../utils';
 
 describe('Simulate', () => {
     let contract: Contract;
